Split queue source binding configurations
