+++ src/Mocha/src/Mocha.Transport.Postgres/Configurations/PostgresQueueDescriptorConfiguration.cs
@@ -72,20 +72,3 @@ public sealed class PostgresQueueDescriptorConfiguration : MessagingConfiguratio
     /// </summary>
     public List<PostgresQueueSourceBindingConfiguration> SourceBindings { get; } = [];
 }
-
-/// <summary>
-/// Configuration for a source topic binding declared from a queue descriptor.
-/// </summary>
-public sealed class PostgresQueueSourceBindingConfiguration
-{
-    /// <summary>
-    /// Gets or sets the source topic address.
-    /// </summary>
-    public Uri Source { get; set; } = null!;
-
-    /// <summary>
-    /// Gets or sets whether the derived subscription is provisioned. <c>null</c> inherits the
-    /// queue's setting, falling back to the transport default.
-    /// </summary>
-    public bool? AutoProvision { get; set; }
-}
